Classification/requirements.txt
torch==1.11.0
torch-cluster==1.5.9
torch-geometric==1.7.0
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-spline-conv==1.2.1
torchvision==0.9.1
matplotlib==3.3.4
numpy==1.20.2
opencv==4.5.3
pandas==1.2.4
pillow==8.2.0
pip==20.0.2
plotly==5.4.0
python==3.8.2
scikit-learn==0.24.2
scipy==1.6.2
shapely==1.8.0
utm==0.7.0

